feat: complete rewrite with fonctionnal test
